+++ docs/fr_FR/Changelog.rst
@@ -8,6 +8,7 @@ ChangeLog
 - Assistant modeles: Mise-à-jour suite renommage des commandes.
 - Tuya 1Ch switch module (TS0001__TZ3000_tqlv4ug4): Mise-à-jour modèle.
 - Modeles pour Xiaomi: Amélioration syntaxe.
+- Page EQ/avancé: Liste pour les types possibles d'attribut.
 
 230106-BETA-2
 -------------